+++ src/main/java/gift/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package gift.service;
+package gift.repository;
 
 import gift.model.Member;
 import java.util.Optional;
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
     Optional<Member> findByEmailAndPassword(String email, String password);
 }
