+++ src/main/java/gift/repository/WishListRepository.java
@@ -1,4 +1,4 @@
-package gift.service;
+package gift.repository;
 
 import gift.model.WishList;
 import java.util.List;
@@ -7,8 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public
-interface WishListRepository extends JpaRepository<WishList, Long> {
+public interface WishListRepository extends JpaRepository<WishList, Long> {
 
     Optional<List<WishList>> findByMemberId(long memberId);
