+++ src/main/java/gift/service/MemberService.java
@@ -3,6 +3,7 @@ package gift.service;
 import gift.exception.RepositoryException;
 import gift.model.Member;
 import gift.model.MemberDTO;
+import gift.repository.MemberRepository;
 import gift.security.JwtUtil;
 import org.springframework.stereotype.Service;
